refactor(cli): share discovery naming helpers
Extract the duplicated endpoint/resource naming helpers used by browser-sniff
and crowd-sniff into a shared discovery package. The golden harness now proves
browser-sniff output and generated CLI artifacts remain byte-identical after
this simplification.
