+++ internal/crowdsniff/specgen.go
@@ -5,6 +5,7 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/mvanhorn/cli-printing-press/v2/internal/discovery"
 	"github.com/mvanhorn/cli-printing-press/v2/internal/spec"
 )
 
@@ -54,7 +55,7 @@ func BuildSpec(name, baseURL string, endpoints []AggregatedEndpoint, auth *Disco
 			endpoint.Params = specParams
 		}
 
-		resourceKey, resourceName := deriveResourceKey(ep.Path)
+		resourceKey, resourceName := discovery.ResourceKey(ep.Path)
 		if resourceKey == "" {
 			resourceKey = "default"
 			resourceName = "default"
@@ -68,9 +69,9 @@ func BuildSpec(name, baseURL string, endpoints []AggregatedEndpoint, auth *Disco
 			resource.Endpoints = make(map[string]spec.Endpoint)
 		}
 
-		endpointName := deriveEndpointName(ep.Method, ep.Path)
+		endpointName := discovery.EndpointName(ep.Method, ep.Path)
 		if _, exists := resource.Endpoints[endpointName]; exists {
-			endpointName = uniqueEndpointName(resource.Endpoints, endpointName)
+			endpointName = discovery.UniqueEndpointName(resource.Endpoints, endpointName)
 		}
 		resource.Endpoints[endpointName] = endpoint
 		resources[resourceKey] = resource
@@ -114,83 +115,6 @@ func ResolveBaseURL(explicit string, candidates []string) string {
 		}
 	}
 	return ""
-}
-
-// --- Helpers adapted from browsersniff/specgen.go ---
-
-func deriveResourceKey(path string) (string, string) {
-	segments := significantSegments(path)
-	if len(segments) == 0 {
-		return "", ""
-	}
-	// Use only the first significant segment as the resource key.
-	// This prevents slashes in resource names which break the generator's
-	// filepath.Join and Cobra Use field.
-	return segments[0], segments[len(segments)-1]
-}
-
-func significantSegments(path string) []string {
-	parts := strings.Split(path, "/")
-	segments := make([]string, 0, len(parts))
-	for _, segment := range parts {
-		segment = strings.TrimSpace(segment)
-		if segment == "" {
-			continue
-		}
-		if strings.HasPrefix(segment, "{") && strings.HasSuffix(segment, "}") {
-			continue
-		}
-		if segment == "api" || isVersionSegment(segment) {
-			continue
-		}
-		segments = append(segments, segment)
-	}
-	return segments
-}
-
-func isVersionSegment(segment string) bool {
-	if len(segment) < 2 || segment[0] != 'v' {
-		return false
-	}
-	for _, r := range segment[1:] {
-		if r < '0' || r > '9' {
-			return false
-		}
-	}
-	return true
-}
-
-func deriveEndpointName(method string, normalizedPath string) string {
-	resource := "endpoint"
-	segments := significantSegments(normalizedPath)
-	if len(segments) > 0 {
-		resource = strings.ReplaceAll(segments[len(segments)-1], "-", "_")
-	}
-
-	switch strings.ToUpper(method) {
-	case "GET":
-		if strings.Contains(normalizedPath, "{") {
-			return "get_" + resource
-		}
-		return "list_" + resource
-	case "POST":
-		return "create_" + resource
-	case "PUT", "PATCH":
-		return "update_" + resource
-	case "DELETE":
-		return "delete_" + resource
-	default:
-		return strings.ToLower(method) + "_" + resource
-	}
-}
-
-func uniqueEndpointName(endpoints map[string]spec.Endpoint, base string) string {
-	for i := 2; ; i++ {
-		name := fmt.Sprintf("%s_%d", base, i)
-		if _, exists := endpoints[name]; !exists {
-			return name
-		}
-	}
 }
 
 // buildAuthConfig converts a DiscoveredAuth into a spec.AuthConfig.
